+++ debug/util/profile-panel-lists.php
@@ -612,7 +612,7 @@ final class PanelListProfilerRunner
             'script' => $this->root . $script,
             'method' => strtoupper($method),
             'uri' => $uri,
-            'host' => 'dev.lanterns.io',
+            'host' => 'foundry.lanterns.io',
             'post' => $post,
             'cookies' => $this->cookies,
             'output' => $outputFile,